requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
PyJWT==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
twilio==9.3.6
urllib3==2.2.3
yarl==1.17.1
